CPC-49: Deprecate `Jmleroux\CircleCi\Api\Pipeline\AllPipelines`
