+++ src/Api/Pipeline/AllPipelines.php
@@ -4,61 +4,23 @@ declare(strict_types=1);
 
 namespace Jmleroux\CircleCi\Api\Pipeline;
 
+use Jmleroux\CircleCi\Api\Project\ProjectPipelines;
 use Jmleroux\CircleCi\Client;
-use Jmleroux\CircleCi\Model\Pipeline;
-use Jmleroux\CircleCi\ValidateClientVersionTrait;
 
 /**
  * Retrieve pipelines of a project, optionally filtered by branch.
  *
- * @author jmleroux <jmleroux.pro@gmail.com>
- * @link   https://circleci.com/docs/api/v2/#operation/listPipelinesForProject
+ * @author     jmleroux <jmleroux.pro@gmail.com>
+ * @link       https://circleci.com/docs/api/v2/#operation/listPipelinesForProject
+ *
+ * @deprecated use Jmleroux\CircleCi\Api\Project\ProjectPipelines
  */
-class AllPipelines
+class AllPipelines extends ProjectPipelines
 {
-    use ValidateClientVersionTrait;
-
-    public function __construct(private readonly Client $client)
+    public function __construct(Client $client)
     {
-        $this->validateClientVersion($client, ['v2']);
-    }
+        @trigger_error(sprintf('%s is deprecated and will be removed in next versions.', AllPipelines::class));
 
-    /**
-     * @return Pipeline[]
-     */
-    public function execute(string $projectSlug, ?int $maxPipelineCount = null, ?string $branch = null): array
-    {
-        $pipelines = [];
-
-        $uri = sprintf('project/%s/pipeline', $projectSlug);
-        $params = [];
-        if (null !== $branch) {
-            $params['branch'] = $branch;
-        }
-
-        $nextPageToken = null;
-        $smallestPipelineNumber = PHP_INT_MAX;
-        if (null === $maxPipelineCount) {
-            $maxPipelineCount = PHP_INT_MAX;
-        }
-
-        do {
-            if (null !== $nextPageToken) {
-                $params['page-token'] = $nextPageToken;
-                unset($params['branch']);
-            }
-
-            $response = json_decode((string)$this->client->get($uri, $params)->getContent());
-            $nextPageToken = $response->next_page_token;
-
-            foreach ($response->items as $item) {
-                $pipeline = Pipeline::createFromApi($item);
-                $lastPipelineNumber = $pipeline->number();
-                $smallestPipelineNumber = min($lastPipelineNumber, $smallestPipelineNumber);
-                $pipelines[] = Pipeline::createFromApi($item);
-            }
-        } while (null !== $nextPageToken && count($pipelines) < $maxPipelineCount);
-
-        return $pipelines;
+        parent::__construct($client);
     }
 }
